requirements.txt
imapclient
google-auth
google-auth-oauthlib
google-auth-httplib2
google-generativeai
geopy
python-dotenv
mysql-connector-python